refactor(solver): extract smt_translate.py for expression translation
